AME-1195 Added code to properly Normalize the Date, Time and duration values to either Date or a Long.  Originally the duration as... methods returned an Integer, But an Integer would not provide enough precision for a Time Duration.
git-svn-id: https://svn.forgerock.org/openam/branches/openam_10.2.0_xacml3_JAS@4762 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d

+++ openam/openam-entitlements/src/main/java/org/forgerock/openam/xacml/v3/Entitlements/FunctionArgument.java
@@ -175,6 +175,7 @@ public abstract class FunctionArgument  {
         }
         return (Double) fArg.getValue(pip);
     }
+
     public Date asTime(XACMLEvalContext pip) throws XACML3EntitlementException {
 
         FunctionArgument fArg = getDataContainer(pip);
@@ -229,23 +230,23 @@ public abstract class FunctionArgument  {
         }
         return (String) fArg.getValue(pip);
     }
-    public Integer asDayTimeDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
+    public Long asDayTimeDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
 
         FunctionArgument fArg = getDataContainer(pip);
 
         if (!fArg.dataType.isType(DataType.Type.XACMLDAYTIMEDURATIONTYPE)) {
             throw new IndeterminateException("type conflict");
         }
-        return (Integer) fArg.getValue(pip);
+        return (Long) fArg.getValue(pip);
     }
-    public Integer asYearMonthDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
+    public Long asYearMonthDuration(XACMLEvalContext pip) throws XACML3EntitlementException {
 
         FunctionArgument fArg = getDataContainer(pip);
 
         if (!fArg.dataType.isType(DataType.Type.XACMLYEARMONTHDURATIONTYPE)) {
             throw new IndeterminateException("type conflict");
         }
-        return (Integer) fArg.getValue(pip);
+        return (Long) fArg.getValue(pip);
     }
     public String asX500Name(XACMLEvalContext pip) throws XACML3EntitlementException {
